requirements.txt
Babel==2.9.1
bottle==0.12.18
bottle-websocket==0.2.9
cffi==1.14.3
docopt==0.6.2
docx-mailmerge==0.5.0
Eel==0.12.4
future==0.18.2
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==0.4.17
Jinja2==2.11.2
lxml==4.6.1
MarkupSafe==1.1.1
num2words==0.5.10
pycparser==2.20
pytz==2021.1
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.1.2
